Moved CSV module into dedicated file
